Add task to migrate site from Locomotive
The task migrates everything that Locomotive CMS tracks, which
unfortunatelly isn’t everything necessary, and tries to change a great
deal of things to Jekyll’s conventions. Nevertheless, this is limitted
and must be thoroughly checked by human(-s). The task can be executed
like this:

    bin/rake migrate \
      DEPLOY_CONFIG=../syslinavinici.cz-wagon/config/deploy.yml \
      DEPLOY_ENV=production \
      ASSET_HOST=alka-engine.s3.eu-central-1.amazonaws.com

Use Rake arguments `-T` and `-D` to find more info.

These are unresolved things:

* Missing assets retrieval. Locomotive CMS can miss a lot of assets
  compared to what it exposes though its API. If there’s something
  linked to `ASSET_HOST` and it’s not available in local media folder
  as set for Netlify CMS config, migration fails with message
  containing the referenced URL. You can then find the asset and copy
  it to the media folder.

* Setting title field. Netlify CMS kind of expects to have a title
  field on collection which serves as item’s label on listings. Find
  the best field yourself and make sure it’s correctly references from
  relation fields.

* Migrating 1:M and M:N relationship values. It seems that Locomotive
  Coal API which is used for migration doesn’t support this at the
  moment. You have to fix this manually.

* Migrating Locomotive-specific Liquid extensions. Again, you have to
  go through everything (pages, layouts, includes and collections)
  manually.

* Transforming simple pages to Markdown. Layouts, pages and snippets
  stay in HTML. Their Front Matter can contain “rich text” which is
  transformed from HTML to Markdown, but not the content. Especially
  for simple pages, it may be viable to move one of the fields from
  the Front Matter to the body, change the file to extension to `.md`
  and enabled it in Netlify CMS (either move to folder collection or
  make it part of file specialized file collection).

* Transforming asset URL in pages, layouts and includes. Nokigiri
  doesn’t handle well Liquid object nor tags, so it’s disabled.

* Localization. Not support at all.

The code is fairly general regardless of these shortcomings. It uses
Jekyll configuration (location of Jekyll config file can be set via
`JEKYLL_CONFIG`) as well as Netlify CMS configuration for file
locations. It tries its best to transforms asset URL in collections,
sets body of Markdown files, finds candidate for the `date` field and
renames it, or adds a new meta fields if no `date` field or candidate
for it is found. Lastly, some prerequisite tasks can run in parallel.