Modify CMS config
Change order of collections to be similar to that of the old
CMS. Remove problematic relationships. Remove categories.

Categories are impossible to model in CMS config and it’s better to
have them as simple lists working in Jekyll than doing something
custom. Moreover, they’re not even used on this website.`